Remove some Leader/Router related features from MTD build
* Remove Leader/Router related APIs from MTD build

* Remove Leader/Router related functions from MTD build
