+++ src/core/thread/mle.hpp
@@ -846,22 +846,6 @@ public:
     void SetAssignLinkQuality(const Mac::ExtAddress aMacAddr, uint8_t aLinkQuality);
 
     /**
-     * This method returns the ROUTER_SELECTION_JITTER value.
-     *
-     * @returns The ROUTER_SELECTION_JITTER value.
-     *
-     */
-    uint8_t GetRouterSelectionJitter(void) const { return mRouterSelectionJitter; }
-
-    /**
-     * This method sets the ROUTER_SELECTION_JITTER value.
-     *
-     * @returns The ROUTER_SELECTION_JITTER value.
-     *
-     */
-    ThreadError SetRouterSelectionJitter(uint8_t aRouterJitter);
-
-    /**
      * This method returns the Child ID portion of an RLOC16.
      *
      * @param[in]  aRloc16  The RLOC16 value.
@@ -1373,10 +1357,6 @@ protected:
 
     Timer mParentRequestTimer;    ///< The timer for driving the Parent Request process.
     Timer mDelayedResponseTimer;  ///< The timer to delay MLE responses.
-
-    uint8_t mRouterSelectionJitter;         ///< The variable to save the assigned jitter value.
-    uint8_t mRouterSelectionJitterTimeout;  ///< The Timeout prior to request/release Router ID.
-
     uint8_t mLastPartitionRouterIdSequence;
     uint32_t mLastPartitionId;
